requirements.txt
certifi==2019.6.16
Django==2.1.7
django-ckeditor==5.6.1
django-cleanup==3.2.0
django-crispy-forms==1.7.2
django-js-asset==1.2.2
django-taggit==1.1.0
olefile==0.46
Pillow==6.0.0
pytz==2018.9
